+++ src/hanami/src/core/processing/cpu/output_processing.h
@@ -62,8 +62,8 @@ processNeuronsOfOutputHexagon(Hexagon* hexagon, uint32_t randomSeed)
     Axon* axon = nullptr;
     AxonBlock* axonBlock = nullptr;
     float weightSum = 0.0f;
-    uint32_t outPos = 0;
-    uint32_t wb = 0;
+    uint64_t outPos = 0;
+    uint64_t wb = 0;
     uint32_t w = 0;
 
     OutputNeuron* out = nullptr;
@@ -76,7 +76,7 @@ processNeuronsOfOutputHexagon(Hexagon* hexagon, uint32_t randomSeed)
     }
 
     assert(outputInterface->weights.size() % outputInterface->outputNeurons.size() == 0);
-    const uint32_t dim = outputInterface->weights.size() / outputInterface->outputNeurons.size();
+    const uint64_t dim = outputInterface->weights.size() / outputInterface->outputNeurons.size();
 
     for (outPos = 0; outPos < outputInterface->outputNeurons.size(); ++outPos) {
         out = &outputInterface->outputNeurons[outPos];
